Align PT origin/od node IDs with their respective zone ID (#672)
* Add `keep_original_ods` to SubAreaAnalysis.post_process

This will attach the original OD matrix entries to the index of the returned matrix. This effectively encodes all the SL
OD information in the returned matrix. The matrix will be significantly larger as no entries will be able to be combined.

* Typo

* Linting

* Minor fixes

* .

* .

* Update docs

* Update tests

* Add custom sqlite3 factory for manual transaction control

* Fix PT skimming on non-contiguous centroids, align origin IDs to zone_ids

Replaces the float max value in the skims with 0.0s as well.

* Fix PT skimming on networks with centroid flow blocked

* Fixes bugs

* Return a namedtuple instead of a module

Update reference files

* Fix manual transaction control, wasn't committing some changes

* Linting

* Update reference files

* Update aequilibrae/project/database_specification/transit/tables/nodes.sql

Co-authored-by: Pedro Camargo <c@margo.co>

---------

Co-authored-by: pveigadecamargo <pveigadecamargo@anl.gov>
Co-authored-by: Pedro Camargo <c@margo.co>